Guess better 64 bit types in fake typedefs.

+++ dependencies/fake_libc_include/_fake_typedefs.h
@@ -118,8 +118,8 @@ typedef signed short int16_t;
 typedef unsigned short uint16_t;
 typedef signed int int32_t;
 typedef unsigned int uint32_t;
-typedef signed long int64_t;
-typedef unsigned long uint64_t;
+typedef signed long long int64_t;
+typedef unsigned long long uint64_t;
 
 
 /* C99 minimum-width integer types */
